refactor : 파일명 풀네임으로 수정
