+++ src/utils/style/mediaQuery.ts
@@ -8,6 +8,7 @@ import {
 
 const breakpoints = [MAX_MOBILE_WIDTH, MAX_TABLET_WIDTH, MAX_DESKTOP_WIDTH];
 
+// media query
 const mq = (cssObject: CSSObject) => {
   const facePaintFn = facepaint(
     breakpoints.map((bp) => `@media (min-width: ${bp}px)`)
